# backend/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Jugueria API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/v1/admin/settings/opening-hours" : {
      "get" : {
        "operationId" : "openingHoursList",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OpeningHourResponse"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Unauthorized. Error codes: auth.unauthenticated",
            "x-error-codes" : [ "auth.unauthenticated" ]
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Forbidden. Error codes: auth.forbidden",
            "x-error-codes" : [ "auth.forbidden" ]
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Internal Server Error. Error codes: common.internal-error",
            "x-error-codes" : [ "common.internal-error" ]
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "opening-hours-controller" ]
      },
      "put" : {
        "operationId" : "openingHoursReplaceAll",
        "parameters" : [ {
          "in" : "header",
          "name" : "If-Match",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/UpdateOpeningHourRequest"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OpeningHourResponse"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Bad Request. Error codes: common.malformed-request, common.validation-failed",
            "x-error-codes" : [ "common.malformed-request", "common.validation-failed" ]
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Unauthorized. Error codes: auth.unauthenticated",
            "x-error-codes" : [ "auth.unauthenticated" ]
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Forbidden. Error codes: auth.forbidden",
            "x-error-codes" : [ "auth.forbidden" ]
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict. Error codes: common.concurrent-modification",
            "x-error-codes" : [ "common.concurrent-modification" ]
          },
          "412" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Precondition Failed. Error codes: common.precondition-failed",
            "x-error-codes" : [ "common.precondition-failed" ]
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Unprocessable Content. Error codes: store.invalid-opening-hours",
            "x-error-codes" : [ "store.invalid-opening-hours" ]
          },
          "428" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Precondition Required. Error codes: common.precondition-required",
            "x-error-codes" : [ "common.precondition-required" ]
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Internal Server Error. Error codes: common.internal-error",
            "x-error-codes" : [ "common.internal-error" ]
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "opening-hours-controller" ]
      }
    },
    "/api/v1/admin/stations" : {
      "get" : {
        "operationId" : "stationList",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StationResponse"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Unauthorized. Error codes: auth.unauthenticated",
            "x-error-codes" : [ "auth.unauthenticated" ]
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Forbidden. Error codes: auth.forbidden",
            "x-error-codes" : [ "auth.forbidden" ]
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Internal Server Error. Error codes: common.internal-error",
            "x-error-codes" : [ "common.internal-error" ]
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "station-controller" ]
      },
      "post" : {
        "operationId" : "stationCreate",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SaveStationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StationResponse"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Bad Request. Error codes: common.malformed-request, common.validation-failed",
            "x-error-codes" : [ "common.malformed-request", "common.validation-failed" ]
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Unauthorized. Error codes: auth.unauthenticated",
            "x-error-codes" : [ "auth.unauthenticated" ]
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Forbidden. Error codes: auth.forbidden",
            "x-error-codes" : [ "auth.forbidden" ]
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict. Error codes: catalog.station-name-already-used",
            "x-error-codes" : [ "catalog.station-name-already-used" ]
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Internal Server Error. Error codes: common.internal-error",
            "x-error-codes" : [ "common.internal-error" ]
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "station-controller" ]
      }
    },
    "/api/v1/admin/stations/{id}" : {
      "put" : {
        "operationId" : "stationRename",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SaveStationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StationResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Bad Request. Error codes: common.malformed-request, common.validation-failed",
            "x-error-codes" : [ "common.malformed-request", "common.validation-failed" ]
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Unauthorized. Error codes: auth.unauthenticated",
            "x-error-codes" : [ "auth.unauthenticated" ]
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Forbidden. Error codes: auth.forbidden",
            "x-error-codes" : [ "auth.forbidden" ]
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Not Found. Error codes: catalog.station-not-found",
            "x-error-codes" : [ "catalog.station-not-found" ]
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict. Error codes: catalog.station-name-already-used, common.concurrent-modification",
            "x-error-codes" : [ "catalog.station-name-already-used", "common.concurrent-modification" ]
          },
          "412" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Precondition Failed. Error codes: common.precondition-failed",
            "x-error-codes" : [ "common.precondition-failed" ]
          },
          "428" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Precondition Required. Error codes: common.precondition-required",
            "x-error-codes" : [ "common.precondition-required" ]
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Internal Server Error. Error codes: common.internal-error",
            "x-error-codes" : [ "common.internal-error" ]
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "station-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ErrorCode" : {
        "type" : "string",
        "description" : "Every error code the API can return; codes are never renamed or reused.",
        "enum" : [ "auth.account-locked", "auth.forbidden", "auth.invalid-credentials", "auth.invalid-refresh-token", "auth.invalid-set-password-token", "auth.unauthenticated", "catalog.category-name-already-used", "catalog.category-not-found", "catalog.currency-mismatch", "catalog.invalid-image", "catalog.invalid-modifier-group", "catalog.invalid-modifier-selection", "catalog.invalid-price", "catalog.invalid-product", "catalog.modifier-group-in-use", "catalog.modifier-group-name-already-used", "catalog.modifier-group-not-found", "catalog.modifier-option-not-found", "catalog.modifier-option-unavailable", "catalog.price-currency-mismatch", "catalog.product-name-already-used", "catalog.product-not-found", "catalog.product-unavailable", "catalog.provider-unavailable", "catalog.station-name-already-used", "catalog.station-not-found", "catalog.unknown-category", "catalog.unknown-modifier-group", "catalog.unknown-station", "common.concurrent-modification", "common.content-too-large", "common.internal-error", "common.malformed-request", "common.method-not-allowed", "common.not-acceptable", "common.not-found", "common.precondition-failed", "common.precondition-required", "common.service-unavailable", "common.unsupported-media-type", "common.validation-failed", "identity.cannot-modify-own-account", "identity.email-already-registered", "identity.invalid-staff-role", "identity.last-active-admin-required", "store.delivery-zone-currency-mismatch", "store.delivery-zone-name-already-used", "store.delivery-zone-not-found", "store.invalid-board-thresholds", "store.invalid-delivery-zone", "store.invalid-opening-hours", "store.invalid-settings-value", "store.reason-code-already-used", "store.reason-not-found" ]
      },
...
      "OpeningHourResponse" : {
        "type" : "object",
        "properties" : {
          "closed" : {
            "type" : "boolean"
          },
          "closesAt" : {
            "type" : "string",
            "format" : "time-local"
          },
          "dayOfWeek" : {
            "type" : "string",
            "enum" : [ "MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY" ]
          },
          "opensAt" : {
            "type" : "string",
            "format" : "time-local"
          }
        }
      },
...
      "Problem" : {
        "type" : "object",
        "additionalProperties" : true,
        "description" : "RFC 9457 Problem Details; extra members depend on the code.",
        "properties" : {
          "code" : {
            "$ref" : "#/components/schemas/ErrorCode"
          },
          "correlationId" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "properties" : {
                "constraint" : {
                  "type" : "string"
                },
                "field" : {
                  "type" : "string"
                },
                "message" : {
                  "type" : "string"
                }
              }
            }
          },
          "instance" : {
            "type" : "string",
            "format" : "uri-reference"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "format" : "uri"
          }
        },
        "required" : [ "code", "status" ]
      },
...
      "SaveStationRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 80,
            "minLength" : 0
          }
        },
        "required" : [ "name" ]
      },
...
      "StationResponse" : {
        "type" : "object",
        "properties" : {
          "defaultStation" : {
            "type" : "boolean"
          },
          "etag" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
...
      "UpdateOpeningHourRequest" : {
        "type" : "object",
        "properties" : {
          "closed" : {
            "type" : "boolean"
          },
          "closesAt" : {
            "type" : "string",
            "format" : "time-local"
          },
          "dayOfWeek" : {
            "type" : "string",
            "enum" : [ "MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY" ]
          },
          "opensAt" : {
            "type" : "string",
            "format" : "time-local"
          }
        },
        "required" : [ "dayOfWeek" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}